backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
pydantic>=2.0.0
python-multipart>=0.0.6
email-validator>=2.0.0
python-dotenv>=1.0.0
google-generativeai>=0.8.3